+++ src/components/TodoContainer.js
@@ -1,0 +1,81 @@
+/* eslint-disable react/destructuring-assignment */
+/* eslint-disable react/no-access-state-in-setstate */
+/* eslint-disable react/state-in-constructor */
+import React, { Component } from 'react';
+import { v4 as uuidv4 } from 'uuid';
+import TodosList from './TodosList';
+import Header from './Header';
+import InputTodo from './InputTodo';
+
+class TodoContainer extends Component {
+  state = {
+    todos: [
+      {
+        id: uuidv4(),
+        title: 'Set up development environment',
+        completed: true,
+      },
+      {
+        id: uuidv4(),
+        title: 'Set up development environment',
+        completed: true,
+      },
+      {
+        id: uuidv4(),
+        title: 'Set up development environment',
+        completed: true,
+      },
+    ],
+  };
+
+  handleChange = (id) => {
+    this.setState((prevState) => ({
+      todos: prevState.todos.map((todo) => {
+        if (todo.id === id) {
+          return {
+            ...todo,
+            completed: !todo.completed,
+          };
+        }
+        return todo;
+      }),
+    }));
+  };
+
+  delTodo = (id) => {
+    this.setState({
+      todos: [
+        ...this.state.todos.filter((todo) => todo.id !== id),
+      ],
+    });
+  };
+
+  addToItem = (title) => {
+    const newTodo = {
+      id: uuidv4(),
+      title,
+      completed: false,
+    };
+    this.setState({
+      todos: [...this.state.todos, newTodo],
+    });
+  }
+
+  render() {
+    return (
+      <div className="container">
+        <div className="inner">
+          <Header />
+          <InputTodo addTodoProps={this.addToItem} />
+          <TodosList
+            todos={this.state.todos}
+            handleChangeProps={this.handleChange}
+            deleteTodoProps={this.delTodo}
+          />
+        </div>
+      </div>
+    );
+  }
+}
+
+export default TodoContainer;
