+++ src/components/TodoItem.js
@@ -1,0 +1,28 @@
+/* eslint-disable react/button-has-type */
+/* eslint-disable react/prop-types */
+/* eslint-disable react/destructuring-assignment */
+import React, { Component } from 'react';
+
+class TodoItem extends Component {
+  handleEditing = () => {
+    console.log('edit mode activated');
+  }
+
+  render() {
+    return (
+      <li>
+        <div onDoubleClick={this.handleEditing}>
+          <input
+            type="checkbox"
+            checked={this.props.todo.completed}
+            onChange={() => this.props.handleChangeProps(this.props.todo.id)}
+          />
+          {this.props.todo.title}
+          <button onClick={() => this.props.deleteTodoProps(this.props.todo.id)}>Delete</button>
+        </div>
+      </li>
+    );
+  }
+}
+
+export default TodoItem;
