+++ src/index.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import ReactDom from 'react-dom';
+// component file
+import TodoContainer from './components/TodoContainer';
+
+ReactDom.render(
+  <React.StrictMode>
+    <TodoContainer />
+  </React.StrictMode>,
+  document.getElementById('root'),
+);
